Move ImageExclusions class to base protocol
